+++ models/account.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Account(BaseModel):
+    api_hash: str
+    api_id: int
+    session_name: str
